requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
flask-marshmallow==1.2.1
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
bcrypt==4.2.1
cryptography==44.0.0
mercadopago==2.3.0
google-generativeai==0.8.4
Pillow==11.1.0
reportlab==4.2.5
flasgger==0.9.7.1
Flask-Limiter==3.9.0
Flask-CORS==4.0.1
Flask-Talisman==1.1.0
python-dotenv==1.0.1
cloudinary==1.41.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
